+++ frontend/src/pages/EmployeeProfile/EmployeeSkillsAndCourses/EmployeeCourses/EmployeeCoursesList/CourseItem/index.ts
@@ -1,0 +1,1 @@
+export { default } from './CourseItemContainer';
