+++ frontend/src/pages/EmployeeProfile/EmployeeSkillsAndCourses/EmployeeCourses/EmployeeCoursesList/CourseItem/types/index.ts
@@ -1,0 +1,14 @@
+import { ConvertedProgress } from 'utils/helpers/convertCourseStatusToProgress';
+
+export interface ICourseItemContainerProps {
+  title: string;
+  status: string;
+}
+
+export interface ICourseItemProps extends ConvertedProgress {
+  title: string;
+  status: string;
+  showCourseInfo?: () => void;
+  hideCourseInfo?: () => void;
+  isShown?: boolean;
+}
