--- frontend/src/pages/EmployeeProfile/EmployeeSkillsAndCourses/EmployeeCourses/EmployeeCoursesList/EmployeeCoursesListContainer.tsx
@@ -1,28 +1,0 @@
-import { FC, useState } from 'react';
-
-import { IEmployeeCoursesList } from 'types/employee';
-
-import EmployeeCoursesList from './EmployeeCourseList';
-
-const EmployeeCoursesListContainer: FC<IEmployeeCoursesList> = ({ courses }) => {
-  const [isShown, setIsShown] = useState(false);
-
-  const showCourseInfo = () => {
-    setIsShown(true);
-  };
-
-  const hideCourseInfo = () => {
-    setIsShown(false);
-  };
-
-  return (
-    <EmployeeCoursesList
-      courses={courses}
-      showCourseInfo={showCourseInfo}
-      hideCourseInfo={hideCourseInfo}
-      isShown={isShown}
-    />
-  );
-};
-
-export default EmployeeCoursesListContainer;
