+++ frontend/src/pages/EmployeeProfile/EmployeeSkillsAndCourses/EmployeeCourses/EmployeeCoursesList/index.ts
@@ -1,1 +1,1 @@
-export { default } from './EmployeeCoursesListContainer';
+export { default } from './EmployeeCoursesList';
